+++ YourEuro/app/src/main/java/com/artexceptionals/youreuro/model/Account.java
@@ -5,9 +5,9 @@ import java.util.Objects;
 public class Account {
 
     String accountName;
-    float balance;
+    double balance;
 
-    public Account(String accountName, float balance) {
+    public Account(String accountName, double balance) {
         this.accountName = accountName;
         this.balance = balance;
     }
@@ -24,11 +24,11 @@ public class Account {
         this.accountName = accountName;
     }
 
-    public float getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(float balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
